requirements.txt
flask==3.1.0
python_magic==0.4.27
python_dotenv==1.1.0
pillow==11.2.1


